add error.log to catch all unchecked exception
