+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix Test loop Command - Alex Thiel
 - Fix RecognizeObjectEvent crash issue
 - Catch System Exception to logger
+- Add error.log to catch all unchecked exception
 
 ## V1.1.4
 - Update - Alex Tan
